requirement.txt
flask
numpy
opencv-python
matplotlib
pytesseract
mysql-connector-python